+++ code/client/snd_local.h
@@ -156,7 +156,7 @@ typedef struct
 bool SNDDMA_Init(void);
 
 // gets the current DMA position
-int SNDDMA_GetDMAPos(void);
+size_t SNDDMA_GetDMAPos(void);
 
 // shutdown the DMA xfer.
 void SNDDMA_Shutdown(void);
